+++ frontend-admin/src/pages/ContentPage/components/ContentContainer/ContentContainer.tsx
@@ -1,9 +1,9 @@
 import Dropdown from '@/components/Dropdown/Dropdown';
-import ContentList from '../ContentList/ContentList';
+import ContentList from './ContentList/ContentList';
 import { dropdownWrapper } from './ContentContainer.styles';
 import useCategoryDropdown from '../../hooks/useCategoryDropdown';
 import useCategoryListQuery from '@/hooks/useCategoryListQuery';
-import ContentAppendButton from '../ContentList/ContentAppendButton/ContentAppendButton';
+import ContentAppendButton from './ContentAppendButton/ContentAppendButton';
 
 const ContentContainer = () => {
   const { data: categoryList } = useCategoryListQuery();
