+++ frontend-admin/src/pages/ContentPage/components/ContentContainer/ContentList/ContentDeleteButton/ContentDeleteButton.styles.ts
