+++ frontend-admin/src/pages/ContentPage/components/ContentContainer/ContentList/DeleteModal/DeleteModal.tsx
