+++ frontend-admin/src/pages/ContentPage/components/ContentContainer/ContentList/OptionCell/OptionCell.styles.ts
