+++ frontend-admin/src/pages/ContentPage/components/ContentContainer/ContentList/QuestionCell/QuestionCell.styles.ts
